+++ src/actions/index.ts
@@ -1,10 +1,12 @@
 import { Config, Action } from "../types";
 import { addInvoice } from "./addInvoice.js";
 import { EVENT } from "../types";
+import { create_invoice } from "./create_invoice.js";
+
 export const actions = async (event: EVENT, options: Config) => {
   switch (event.queryStringParameters?.action) {
-    case "add_invoice":
-      return await addInvoice(event, options);
+    case "create_invoice":
+      return await create_invoice(event, options);
     default:
       throw "Route not found";
   }
@@ -12,8 +14,8 @@ export const actions = async (event: EVENT, options: Config) => {
 
 export const actionsList: Action[] = [
   {
-    action: "add_invoice",
-    name: "add invoice",
-    description: "add invoice ..",
+    action: "create_invoice",
+    name: "create invoice",
+    description: "create invoice ..",
   },
 ];