+++ changelog.md
@@ -4,6 +4,12 @@
 
 ### Added
 
+- [./src/commands/tax] new Command: Tax @maramalshen
+- [./src/commands/product] implement tax key in the product @maramalshen
+- [./src/commands/measureunit_family.ts] new Command: Measure Unit Family @maramalshen
+- [./src/commands/product] handle product.track_quantity as product.frozen @maramalshen
+- [./src/actions/create_invoice] create invoice in Qoyod @maramalshen
+
 ### Changed
 
 ### Fixed